# Speed up generation of output