--- a/package.json
+++ b/package.json
@@ -44,6 +44,7 @@
 	],
 	"dependencies": {
 		"ansi-escapes": "^6.0.0",
+		"ansi-tokenize": "github:AlCalzone/ansi-tokenize",
 		"auto-bind": "^5.0.1",
 		"chalk": "^5.2.0",
 		"cli-boxes": "^3.0.0",