--- a/test/overflow.tsx
+++ b/test/overflow.tsx
@@ -502,19 +502,22 @@ test('nested overflow', t => {
 	t.is(output, 'AA\nBB\nXXXX\nYYYY\n');
 });
 
-test('out of bounds writes are truncated and do not crash', t => {
-	// To trigger this reliably, render a box with a border that is at least 2 columns too wide for the terminal
+// See https://github.com/vadimdemedes/ink/pull/564#issuecomment-1637022742
+test('out of bounds writes do not crash', t => {
 	const output = renderToString(
-		<Box width={12} height={10} borderStyle={'round'} flexDirection="row"></Box>,
+		<Box width={12} height={10} borderStyle="round" />,
 		{columns: 10}
 	);
+
 	const expected = boxen('', {
-		width: 11,
+		width: 12,
 		height: 10,
 		borderStyle: 'round'
 	})
 		.split('\n')
-		.map(line => line.slice(0, 10).trimEnd())
+		.map((line, index) => {
+			return index === 0 || index === 9 ? line : line.slice(0, 10) + line[11];
+		})
 		.join('\n');
 
 	t.is(output, expected);