This is a proof of concept for a different rendering approach that does not rely on `sliceAnsi` as heavily.

I tried this with my test component from #560, but with 50 spinners. Without this change, 50 spinners put one core of my CPU at 100%, and rendering lags heavily. With this change, I'm looking at ~40% utilization of one core, which is still not ideal, but renders fluently.

The output for 10s of profiling now looks like this, is idle half the time and most of the work is done in `wasm`, which is probably from `yoga` internals:
![grafik](https://user-images.githubusercontent.com/17641229/226215080-f92d2d84-de67-4f71-92ab-7c95a1627099.png)

fixes: https://github.com/vadimdemedes/ink/issues/560

This will need some more testing with emojis and clipping, but I wanted to get this out for some early feedback.