@AlCalzone 2023-03-20
Ok, looks like emojis and some full-width characters are causing test failures, I'll look into that.

And there's probably an ANSI code in a different position than before - I don't see a visual difference here:
![grafik](https://user-images.githubusercontent.com/17641229/226435027-bb23369f-e083-4382-9008-d55cd2030679.png)