@AlCalzone 2023-04-05
Tests are passing too now. Just for fun, I ran `time npm run benchmark benchmark/simple` to compare.
**master**
```
real    0m11,412s
user    0m13,342s
sys     0m0,604s
```

**this PR**
```
real    0m8,208s
user    0m9,051s
sys     0m0,330s
```

So roughly 32% faster rendering for even simple outputs.