@vadimdemedes 2023-07-16
Thanks @AlCalzone, that was really helpful! Here's what I've found.

I took your `Box` example and rendered it in the terminal.

```jsx
import React from 'react';
import {Box, render} from 'ink';

function Test() {
	return <Box width={12} height={10} borderStyle={'round'}></Box>;
}

process.stdout.columns = 10;
render(<Test />, {debug: true});
```

<img width="327" alt="CleanShot 2023-07-16 at 11 13 06@2x" src="https://github.com/vadimdemedes/ink/assets/697676/44833ca7-1e02-4d54-8ab5-af851beff439">

The right border is rendered in the wrong place, even though there are correct X and Y coordinates for all sides:

- Top border has X = 0, Y = 0
- Left border has X = 0, Y = 1
- Right border has X = 11, Y = 1
- Bottom border has X = 0, Y = 9

Here's where each line is written to output:

https://github.com/vadimdemedes/ink/blob/8a0476024cf997e0dbc02e493ad3972349b02474/src/output.ts#L187-L190

I added this `console.log` to see what exactly gets written and where:

```js
console.log({
	// X coordinate of the `line` to write to output
	x,
	// Existing output before X coordinate
	before: sliceAnsi(currentLine, 0, x),
	beforeWidth: sliceAnsi(currentLine, 0, x).length,
	// New text to write to output
	line,
	lineWidth: line.length,
	// Existing output after X coordinate + width of the `line`
	after: sliceAnsi(currentLine, x + width),
	afterWidth: sliceAnsi(currentLine, x + width).length
});
```

Here's the output for one of the lines for writing the right border:

```
{
  x: 11,
  before: '│         ',
  beforeWidth: 10,
  line: '│',
  lineWidth: 1,
  after: '',
  afterWidth: 0
}
```

Normally, `sliceAnsi(currentLine, 0, x)` should return a string where `length === x`, but here `beforeWidth` equals 10, not 11. The reason is `currentLine` comes from the `output` variable, which is initialized with strings that are as wide as the terminal (in our case terminal width is 10).

https://github.com/vadimdemedes/ink/blob/8a0476024cf997e0dbc02e493ad3972349b02474/src/output.ts#L174

https://github.com/vadimdemedes/ink/blob/8a0476024cf997e0dbc02e493ad3972349b02474/src/output.ts#L97-L99

This creates a problem where `output` contains lines with the length of 10, so the index of the last character in each line is 9. At the same time, we're trying to write a character at index 11, which results in invalid output where character at index 10 is missing. As a result, right border is placed at X = 10, not X = 11.

It doesn't cause a crash in the current version of Ink, because `sliceAnsi` is used to extract string ranges, which doesn't fail if you run `sliceAnsi(str, 0, 11)` where `str.length === 10`. However, your PR no longer uses `sliceAnsi` and is sensitive to `undefined` items.

To sum up, this is an existing issue of Ink with content larger than the terminal and should be fixed separately from this PR. To avoid regressions in this PR though, we need to skip `undefined` array items to keep the rendering behavior the same as it is now.