@vadimdemedes 2023-07-16
Pushed a commit to keep existing behavior with out of bounds writes.