@vadimdemedes 2023-07-16
CI is green, merging!

Thank you for this incredible performance boost, @AlCalzone!